Add files via upload
This version has gone through significant changes, and it also has been moved to another repository due to issues with the first one. The changes are as follows: a save_csv function has been added to save the output of app.py, and a save_qualifying_loans function has been created to ask the user if they wish to save the output of the function to a folder called output within data. Testing has also been added to the app."